Move Policy tests into the new structure
