+++ tests/protocols/gmpv208/entities/policies/test_delete_policy.py
@@ -16,35 +16,27 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import RequiredArgument
+from gvm.errors import GvmError
 
 
-class GmpCreatePolicyTestCase:
-    def test_create_policy(self):
-        self.gmp.create_policy('foo')
-
-        self.connection.send.has_been_called_with(
-            '<create_config>'
-            '<copy>085569ce-73ed-11df-83c3-002264764cea</copy>'
-            '<name>foo</name>'
-            '<usage_type>policy</usage_type>'
-            '</create_config>'
-        )
-
-    def test_create_with_policy_id(self):
-        self.gmp.create_policy('foo', policy_id='p1')
+class GmpDeletePolicyTestMixin:
+    def test_delete(self):
+        self.gmp.delete_policy('a1')
 
         self.connection.send.has_been_called_with(
-            '<create_config>'
-            '<copy>p1</copy>'
-            '<name>foo</name>'
-            '<usage_type>policy</usage_type>'
-            '</create_config>'
+            '<delete_config config_id="a1" ultimate="0"/>'
         )
 
-    def test_missing_name(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.create_policy(policy_id='c1', name=None)
+    def test_delete_ultimate(self):
+        self.gmp.delete_policy('a1', ultimate=True)
 
-        with self.assertRaises(RequiredArgument):
-            self.gmp.create_policy(policy_id='c1', name='')
+        self.connection.send.has_been_called_with(
+            '<delete_config config_id="a1" ultimate="1"/>'
+        )
+
+    def test_missing_config_id(self):
+        with self.assertRaises(GvmError):
+            self.gmp.delete_policy(None)
+
+        with self.assertRaises(GvmError):
+            self.gmp.delete_policy('')
