+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -27,28 +27,9 @@ from .test_modify_permission import GmpModifyPermissionTestCase
 from .test_modify_report_format import GmpModifyReportFormatTestCase
 from .test_modify_tag import GmpModifyTagTestCase
 from .test_get_feed import GmpGetFeedTestCase
-from .test_clone_policy import GmpClonePolicyTestCase
 from .test_create_credential import GmpCreateCredentialTestCase
-from .test_create_policy import GmpCreatePolicyTestCase
 from .test_create_scanner import GmpCreateScannerTestCase
-from .test_delete_policy import GmpDeletePolicyTestCase
-from .test_get_policies import GmpGetPoliciesTestCase
-from .test_get_policy import GmpGetPolicyTestCase
 from .test_modify_credential import GmpModifyCredentialTestCase
-from .test_modify_policy_set_comment import GmpModifyPolicySetCommentTestCase
-from .test_modify_policy_set_family_selection import (
-    GmpModifyPolicySetFamilySelectionTestCase,
-)
-from .test_modify_policy_set_name import GmpModifyPolicySetNameTestCase
-from .test_modify_policy_set_nvt_preference import (
-    GmpModifyPolicySetNvtPreferenceTestCase,
-)
-from .test_modify_policy_set_nvt_selection import (
-    GmpModifyPolicySetNvtSelectionTestCase,
-)
-from .test_modify_policy_set_scanner_preference import (
-    GmpModifyPolicySetScannerPreferenceTestCase,
-)
 from .test_modify_scanner import GmpModifyScannerTestCase
 from .test_modify_ticket import GmpModifyTicketTestCase
 from .test_clone_ticket import GmpCloneTicketTestCase
